fix(backend): consolidate duplicate error handlers into single canonical handler
INFRA-001 — Orphaned Root-Level src/ Directory Duplicates Frontend Code
Fixes #621

INFRA-002 — Conflicting Jest Config Files in Backend
Fixes #622

INFRA-003 — Two Competing Error Handlers in Backend
Fixes #623

INFRA-005 — Working Tree Has 28 Uncommitted Modified Files
Fixes #625
