+++ backend/src/__tests__/escrow.release.routes.test.ts
@@ -4,7 +4,7 @@ import request from "supertest";
 import * as StellarSdk from "@stellar/stellar-sdk";
 import { createEscrowReleaseRouter } from "../routes/escrow.release.routes";
 import { AuthService } from "../services/auth.service";
-import { errorHandler } from "../errors/errorHandler";
+import { errorHandler } from "../middleware/errorHandler";
 
 jest.mock("../services/auth.service", () => ({
   AuthService: {
